+++ src/models/generateddata.js
@@ -3,7 +3,7 @@ const LocalizationConfig = require('./localizationconfig');
 const PageConfig = require('./pageconfig');
 const PageSet = require('./pageset');
 const PageSetsBuilder = require('../commands/build/pagesetsbuilder');
-const PagePartial = require('./pagetemplate');
+const PagePartial = require('./pagepartial');
 
 /**
  * GeneratedData is a representation of the data generated by Jambo. It deals with
